fix: resolved next js route issue
