+++ nextjs/src/app/layout.tsx
@@ -13,7 +13,7 @@ import SessionCheck from '@/features/auth/components/SessionCheck'
 import { SessionSyncer } from '@/features/auth/components/sessionSyncher'
 import StoreProvider from './StoreProvider'
 import { Toaster } from '@/components/ui/sonner'
-import { authOptions } from './api/auth/[...nextauth]/route'
+import { authOptions } from '@/utils/authOptions'
 import { cn } from '@/lib/utils'
 import { cookies } from 'next/headers'
 import { fontVariables } from '@/lib/font'
