--- nextjs/src/lib/auth.ts
@@ -1,4 +1,0 @@
-import NextAuth from 'next-auth'
-import authConfig from './auth.config'
-
-export const { auth, handlers, signOut, signIn } = NextAuth(authConfig)
